requirements.txt
flask
newspaper3k
